requirements.txt
Django==3.2.18
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==3.10.0